+++ NEWS.md
@@ -1,5 +1,5 @@
 # eph 0.2.0
-* enhace get_microdata. Now downloads multiple datasets
+* enhace get_microdata. Now downloads multiple datasets and allows to pre-select variables
 * add skip_if_offline and some skip_on_cran for time-consuming tests
 * add stop if there is no internet connection for get_microdata
 * add vignettes